[#4] Refactor package
