+++ src/main/java/com/depromeet/um/api/domain/model/MongoSequence.java
@@ -1,4 +1,4 @@
-package com.depromeet.um.api.model;
+package com.depromeet.um.api.domain.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
