+++ src/main/java/com/depromeet/um/api/domain/repository/MessageRepository.java
@@ -1,6 +1,6 @@
-package com.depromeet.um.api.repository;
+package com.depromeet.um.api.domain.repository;
 
-import com.depromeet.um.api.model.Message;
+import com.depromeet.um.api.domain.model.Message;
 import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.repository.MongoRepository;
