+++ src/main/java/com/depromeet/um/api/domain/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.depromeet.um.api.repository;
+package com.depromeet.um.api.domain.repository;
 
-import com.depromeet.um.api.model.User;
+import com.depromeet.um.api.domain.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
