+++ src/main/java/com/depromeet/um/api/dto/ChatRoomType.java
@@ -1,0 +1,6 @@
+package com.depromeet.um.api.dto;
+
+public enum ChatRoomType {
+    DIRECT,
+    GROUP
+}
