--- src/main/java/com/depromeet/um/api/repository/ChatRoomRepository.java
@@ -1,9 +1,0 @@
-package com.depromeet.um.api.repository;
-
-import com.depromeet.um.api.model.ChatRoom;
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ChatRoomRepository extends MongoRepository<ChatRoom, Long> {
-}
