+++ src/main/java/com/depromeet/um/api/service/AppleVerifyService.java
@@ -11,7 +11,7 @@ import com.auth0.jwt.exceptions.InvalidClaimException;
 import com.auth0.jwt.exceptions.JWTDecodeException;
 import com.auth0.jwt.exceptions.TokenExpiredException;
 import com.auth0.jwt.interfaces.DecodedJWT;
-import com.depromeet.um.api.model.AppleUser;
+import com.depromeet.um.api.domain.model.AppleUser;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.stereotype.Service;
