+++ src/main/java/com/depromeet/um/api/service/KakaoAccountService.java
@@ -2,13 +2,15 @@ package com.depromeet.um.api.service;
 
 import com.depromeet.um.api.auth.TokenProvider;
 import com.depromeet.um.api.auth.UserSession;
+import com.depromeet.um.api.domain.KakaoUserService;
+import com.depromeet.um.api.domain.UserService;
 import com.depromeet.um.api.dto.KakaoLoginRequest;
 import com.depromeet.um.api.dto.KakaoRegisterRequest;
 import com.depromeet.um.api.dto.LoginRequest;
 import com.depromeet.um.api.dto.LoginType;
 import com.depromeet.um.api.dto.RegisterRequest;
-import com.depromeet.um.api.model.KakaoUser;
-import com.depromeet.um.api.model.User;
+import com.depromeet.um.api.domain.model.KakaoUser;
+import com.depromeet.um.api.domain.model.User;
 import com.depromeet.um.api.util.StringUtils;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.stereotype.Service;
